Scraping tools for OECD related information.

Scraping tools for OECD related information.

oecd_members()Scrapes member information of OECD countries from Wikipedia and returns it in the form of a dataframe. This dataframe contains member information such as accession dates, country names, etc.

oecd_member_metrics() Scrapes member information of OECD countries from Wikipedia and returns it in the form of a dataframe. This dataframe contains economic data of member countries. Not to be used for analysis purposes.

oecd_secretary() scrapes the table of previous OECD Secretary-Generals from Wikipedia and returns a dataframe with their details.

oecd_members(code = FALSE)

oecd_member_metrics()

oecd_secretary()

Arguments

code
Logical. If TRUE, prints the code to create the vector of OECD countries to the console. FALSE by default, returns the whole dataframe.

Source

https://en.wikipedia.org/wiki/OECD

Details

This family of functions returns information related to the OECD such as the member States, the name and tenure period of the secretary- generals of the OECD, etc. Data should be used carefully and not for analytical purposes.

Examples

library(magrittr)
oecd_members()

#> # A tibble: 38 × 8
#>    ID    Country        Application Negotiations Invitation Membership `Geographic loc…
#>    <chr> <chr>          <date>      <date>       <date>     <date>     <chr>           
#>  1 AUS   Australia      NA          NA           NA         1971-06-07 Oceania         
#>  2 AUT   Austria        NA          NA           NA         1961-09-29 Europe          
#>  3 BEL   Belgium        NA          NA           NA         1961-09-13 Europe          
#>  4 CAN   Canada         NA          NA           NA         1961-04-10 North America   
#>  5 CHL   Chile          NA          2007-05-16   2009-12-15 2010-05-07 South America   
#>  6 COL   Colombia       2011-01-24  2013-05-30   2018-05-25 2020-04-28 South America   
#>  7 CRI   Costa Rica     2015-04-09  NA           2020-05-15 2021-05-25 North America   
#>  8 CZE   Czech Republic NA          1994-06-08   1995-11-24 1995-12-21 Europe          
#>  9 DNK   Denmark        NA          NA           NA         1961-05-30 Europe          
#> 10 EST   Estonia        NA          2007-05-16   2010-05-10 2010-12-09 Europe          
#> # … with 28 more rows, and 1 more variable: Notes <chr>

library(magrittr)
oecd_member_metrics()

#> # A tibble: 39 × 15
#>    ID    Country        `Area(km2)2017` `Population 2017` `GDP (PPP) (Intl. $)2…
#>    <chr> <chr>                    <dbl>             <dbl>                  <dbl>
#>  1 AUS   Australia              7741220          24598933          1192065505301
#>  2 AUT   Austria                  83879           8809212           461582926400
#>  3 BEL   Belgium                  30530          11372068           544041974958
#>  4 CAN   Canada                 9984670          36708083          1714447151944
#>  5 CHL   Chile                   756096          18054726           444777637169
#>  6 COL   Colombia               1141748          48901066           709420539907
#>  7 CRI   Costa Rica               51100           5048992            84031346801
#>  8 CZE   Czech Republic           78870          10591323           384753663283
#>  9 DNK   Denmark                  42922           5769603           296350723354
#> 10 EST   Estonia                  45230           1315480            41756008089
#> # … with 29 more rows, and 10 more variables:
#> #   GDP (PPP)per capita(Intl. $)2017 <dbl>,
#> #   Incomeinequality 2008-2016(latest available) <dbl>, HDI2019 <dbl>,
#> #   FSI2019 <dbl>, RLI2020 <dbl>, CPI2019 <dbl>, IEF2020 <dbl>, GPI2019 <dbl>,
#> #   WPFI2019 <dbl>, DI2019 <dbl>

library(magrittr)
oecd_secretary()

#> # A tibble: 8 × 5
#>   ID    `Secretary-General`       Beg        End        `Country of origin`
#>   <chr> <chr>                     <date>     <date>     <chr>              
#> 1 1     Thorkil Kristensen        1961-09-30 1969-09-30 Denmark            
#> 2 2     Emiel van Lennep          1969-10-01 1984-09-30 Netherlands        
#> 3 3     Jean-Claude Paye          1984-10-01 1994-09-30 France             
#> 4 —     Staffan Sohlman (interim) 1994-10-01 1994-11-01 Sweden             
#> 5 3     Jean-Claude Paye          1994-11-01 1996-05-31 France             
#> 6 4     Donald Johnston           1996-06-01 2006-05-31 Canada             
#> 7 5     José Ángel Gurría         2006-06-01 2021-05-31 Mexico             
#> 8 6     Mathias Cormann           2021-06-01 2021-09-02 Australia